requirements.txt
langchain
openai
huggingface_hub
python-dotenv
streamlit
pygame
langchain_openai
streamlit-audiorec
google-generativeai
ipython
gtts
SpeechRecognition
pyaudio
openai-whisper